// frontend/one-nation-auto-loan-ff/src/app/interfaces/searchParams.ts
export interface SearchParams {
  firstName: string;
  lastName: string;
  email: string;
  status: string;
  dealStatus: string;
  referralCode: string;
  tempDealerId: string;
  display: string;
  completionStartDate: Date;
  completionEndDate: Date;
  createdStartDate: Date;
  createdEndDate: Date;
}